Net: Store all the LocalSockets in a HashMap
Like we already do for TCPSockets and UDPSockets. For LocalSockets,
we're using the address of the underlying open file description as a
key into the map. This should work, because the file is opened each
time a socket is connected or bound to the address, so there are as
many different file descriptors pointing to the underlying file on
the file system as there are LocalSockets using this address.